Rally: Fetch a single bug, if needed
As rally just searches for the numbers when using userfriendly formatted
IDs for searching it's possible to have multiple search results when
only one was expected. This change tries to guess based on the executed
query if a single bug was expected, and if so, sees if showing a single
bug can be satisfied from the returned search results.