bz-list: get bug ID via text properties instead of text parsing
Instead of trying to extract the bug ID from the first element each line
now contains a bz-bug-id text property containing the UUID of the
bug. This makes the whole thing more robust, no longer requires an ID
entry at the beginning of the list (or even in the list at all), and
saves one API call for Rally.