rally: add project functionality, fix POST data
Add functionality for querying and creating projects.

This also correctly handle data on POST requests - project was the
first endpoint to actually use POST.

We expect different alist formats for POST and GET requests, and have
been incorrectly nesting the list deeper for POST requests. With that
fixed implementation of write functions is now possible, so mark
the backend as having write support as well.